requirements.txt
flet
python-dotenv
psycopg2-binary
asyncpg
